pyproject.toml
[tool.poetry]
name = "flask.app"
version = "0.1.0"
description = ""
authors = ["denys <denysburimov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Login = "^0.5.0"
click = "^8.1.0"
email-validator = "^1.1.3"
Flask-WTF = "^1.0.0"
WTForms = "^3.0.1"
itsdangerous = "^2.1.2"
Jinja2 = "^3.1.1"
MarkupSafe = "^2.1.1"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.32"
Werkzeug = "2.0.0"
Flask-Migrate = "^3.1.0"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
PyQRCode = "^1.2.1"
onetimepass = "^1.0.1"
timecop = "^0.5.0-alpha.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
flake8 = "^4.0.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
